requirements.txt
zstandard
requests
python-dateutil
tqdm
tldextract
beautifulsoup4
newspaper3k
htmlmin
lm_dataformat
jsonlines
datasketch
colorama
cutie
sqlalchemy
tqdm-multiprocess
base36
alembic
cassandra-driver